introduce styles a lot of improvements regarding fonts rendering introduce TextBlock control huge amount of improvements and bug fixes

+++ Adamantium/Adamantium.Engine.Effects/EffectCompilerInternal.cs
@@ -1220,7 +1220,7 @@ namespace Adamantium.Engine.Effects
                 var parameterBuffer = new EffectData.ConstantBuffer()
                 {
                     Name = description.Name,
-                    Size = (int)description.Size,
+                    Size = description.Size,
                     Slot = description.SlotIndex,
                     Parameters = new List<EffectData.ValueTypeParameter>(),
                 };
